+++ app/Views/DataTable/data_produk_hukum.php
@@ -24,17 +24,21 @@
                             <i class="fa fa-download"></i>
                             Export
                         </a>
-                        <a href="/ProdukHukum/add" class="btn btn-primary" style="float: right">
-                            <i class="fa fa-plus"></i>
-                            Add Produk Hukum
-                        </a>
+                        <?php if (session()->get('role_id') != 1) { ?>
+                            <a href="/ProdukHukum/add" class="btn btn-primary" style="float: right">
+                                <i class="fa fa-plus"></i>
+                                Tambah Produk Hukum
+                            </a>
+                        <?php } ?>
                         <ul class="nav nav-tabs bar_tabs" style="margin-top: 100px;" id="myTab" role="tablist">
                             <li class="nav-item">
                                 <a class="nav-link active" id="tervalidasi-tab" data-toggle="tab" href="#tervalidasi" role="tab" aria-controls="tervalidasi" aria-selected="true">Tervalidasi</a>
                             </li>
-                            <li class="nav-item">
-                                <a class="nav-link" id="blmTervalidasi-tab" data-toggle="tab" href="#blmTervalidasi" role="tab" aria-controls="blmTervalidasi" aria-selected="false">Belum Tervalidasi</a>
-                            </li>
+                            <?php if ($akun['role_id'] != 1) { ?>
+                                <li class="nav-item">
+                                    <a class="nav-link" id="blmTervalidasi-tab" data-toggle="tab" href="#blmTervalidasi" role="tab" aria-controls="blmTervalidasi" aria-selected="false">Belum Tervalidasi</a>
+                                </li>
+                            <?php } ?>
                         </ul>
                         <div class="tab-content" id="myTabContent">
                             <div class="tab-pane fade show active" id="tervalidasi" role="tabpanel" aria-labelledby="tervalidasi-tab">
@@ -51,7 +55,7 @@
                                                 <th class="text-center">Kategori</th>
                                                 <th class="text-center">Tahun</th>
                                                 <th class="text-center">Keterangan</th>
-                                                <th class="text-center">Status</th>
+                                                <th class="text-center">Status Berlaku</th>
                                                 <th class="text-center">Actions</th>
                                             </tr>
                                         </thead>
@@ -69,12 +73,27 @@
                                                     <td><?= $val['nama_kategori']; ?></td>
                                                     <td><?= $val['tahun']; ?></td>
                                                     <td><?= $val['keterangan']; ?></td>
-                                                    <td><?= $val['status']; ?></td>
+                                                    <td>
+                                                        <?php if ($akun['role_id'] == 3) { ?>
+                                                            <center><input type="checkbox" class="js-switch ganti-status" data-id="<?= $val['id_produk']; ?>" <?= ($val['status'] == "Berlaku") ? 'checked' : ''; ?> /></center>
+                                                        <?php } else { ?>
+                                                            <?= $val['status']; ?>
+                                                        <?php } ?>
+                                                    </td>
                                                     <td>
                                                         <center>
                                                             <a href="/ProdukHukum/detail/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-primary"><i class="fa fa-eye"></i></a>
-                                                            <a href="/ProdukHukum/update/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-success"><i class="fa fa-pencil"></i></a>
-                                                            <a href="/ProdukHukum/delete/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-danger tombol-hapus"><i class="fa fa-trash"></i></a>
+                                                            <?php if ($akun['role_id'] == 3) { ?>
+                                                                <a href="/ProdukHukum/delete/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-danger tombol-hapus"><i class="fa fa-trash"></i></a>
+                                                                <div class="dropdown">
+                                                                    <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                                                        <i class="fa fa-gear"></i>
+                                                                    </button>
+                                                                    <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+                                                                        <a class="dropdown-item" href="/ProdukHukum/unvalidation/<?= md5($val['id_produk']); ?>">Batal Validasi</a>
+                                                                    </div>
+                                                                </div>
+                                                            <?php } ?>
                                                         </center>
                                                     </td>
                                                 </tr>
@@ -83,66 +102,68 @@
                                     </table>
                                 </div>
                             </div>
-                            <div class="tab-pane fade" id="blmTervalidasi" role="tabpanel" aria-labelledby="blmTervalidasi-tab">
-                                <div class="card-box table-responsive">
-                                    <form action="/ProdukHukum/validasi_prohum" method="POST">
-                                        <input type="hidden" name="id_unit" value="<?= $request->uri->getSegment(2); ?>">
+                            <?php if ($akun['role_id'] != 1) { ?>
+                                <div class="tab-pane fade" id="blmTervalidasi" role="tabpanel" aria-labelledby="blmTervalidasi-tab">
+                                    <div class="card-box table-responsive">
+                                        <form action="/ProdukHukum/validasi_prohum" method="POST">
+                                            <input type="hidden" name="id_unit" value="<?= $request->uri->getSegment(2); ?>">
 
-                                        <table id="datatable-checkbox" class="table table-striped table-bordered bulk_action" style="width:100%">
-                                            <div class="flash-data" data-flashdata="<?= session()->getFlashdata('prohum'); ?>"></div>
-                                            <br><br><br>
-                                            <thead>
-                                                <tr>
-                                                    <?php if ($akun['role_id'] == 1) { ?>
-                                                        <th>
-                                                            <center><i class="fa fa-check-square-o"></i></center>
-                                                        </th>
-                                                    <?php } ?>
-                                                    <th class="text-center">Unit</th>
-                                                    <th class="text-center">No</th>
-                                                    <th class="text-center">Judul</th>
-                                                    <th class="text-center">Tentang</th>
-                                                    <th class="text-center">Kategori</th>
-                                                    <th class="text-center">Tahun</th>
-                                                    <th class="text-center">Keterangan</th>
-                                                    <th class="text-center">Status</th>
-                                                    <th class="text-center">Actions</th>
-                                                </tr>
-                                            </thead>
-
-                                            <tbody>
-                                                <?php foreach ($prohum_blmValid as $val) { ?>
+                                            <table id="datatable-checkbox" class="table table-striped table-bordered bulk_action" style="width:100%">
+                                                <div class="flash-data" data-flashdata="<?= session()->getFlashdata('prohum'); ?>"></div>
+                                                <br><br><br>
+                                                <thead>
                                                     <tr>
-                                                        <?php if ($akun['role_id'] == 1) { ?>
-                                                            <td class="a-center ">
-                                                                <input type="checkbox" class="flat" name="id[]" value="<?= $val['id_produk']; ?>">
-                                                            </td>
+                                                        <?php if ($akun['role_id'] == 3) { ?>
+                                                            <th>
+                                                                <center><i class="fa fa-check-square-o"></i></center>
+                                                            </th>
                                                         <?php } ?>
-                                                        <td><?= $val['nama_unit']; ?></td>
-                                                        <td><?= $val['no']; ?></td>
-                                                        <td><?= $val['judul']; ?></td>
-                                                        <td><?= (str_word_count($val['nama_tentang'])) > 10 ? substr($val['nama_tentang'], 0, 100) . " [..]" : $val['nama_tentang']; ?></td>
-                                                        <td><?= $val['nama_kategori']; ?></td>
-                                                        <td><?= $val['tahun']; ?></td>
-                                                        <td><?= $val['keterangan']; ?></td>
-                                                        <td><?= $val['status']; ?></td>
-                                                        <td>
-                                                            <center>
-                                                                <a href="/ProdukHukum/detail/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-primary"><i class="fa fa-eye"></i></a>
-                                                                <a href="/ProdukHukum/update/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-success"><i class="fa fa-pencil"></i></a>
-                                                                <a href="/ProdukHukum/delete/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-danger tombol-hapus"><i class="fa fa-trash"></i></a>
-                                                            </center>
-                                                        </td>
+                                                        <th class="text-center">Unit</th>
+                                                        <th class="text-center">No</th>
+                                                        <th class="text-center">Judul</th>
+                                                        <th class="text-center">Tentang</th>
+                                                        <th class="text-center">Kategori</th>
+                                                        <th class="text-center">Tahun</th>
+                                                        <th class="text-center">Keterangan</th>
+                                                        <th class="text-center">Actions</th>
                                                     </tr>
-                                                <?php } ?>
-                                            </tbody>
-                                        </table>
-                                        <?php if ($akun['role_id'] == 1) { ?>
-                                            <button type="submit" class="btn btn-primary mt-5"><i class="fa fa-check-square-o"></i> Validasi</button>
-                                        <?php } ?>
-                                    </form>
+                                                </thead>
+
+                                                <tbody>
+                                                    <?php foreach ($prohum_blmValid as $val) { ?>
+                                                        <tr>
+                                                            <?php if ($akun['role_id'] == 3) { ?>
+                                                                <td class="a-center ">
+                                                                    <input type="checkbox" class="flat" name="id[]" value="<?= $val['id_produk']; ?>">
+                                                                </td>
+                                                            <?php } ?>
+                                                            <td><?= $val['nama_unit']; ?></td>
+                                                            <td><?= $val['no']; ?></td>
+                                                            <td><?= $val['judul']; ?></td>
+                                                            <td><?= (str_word_count($val['nama_tentang'])) > 10 ? substr($val['nama_tentang'], 0, 100) . " [..]" : $val['nama_tentang']; ?></td>
+                                                            <td><?= $val['nama_kategori']; ?></td>
+                                                            <td><?= $val['tahun']; ?></td>
+                                                            <td><?= $val['keterangan']; ?></td>
+                                                            <td>
+                                                                <center>
+                                                                    <a href="/ProdukHukum/detail/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-primary"><i class="fa fa-eye"></i></a>
+                                                                    <?php if ($akun['role_id'] == 2) { ?>
+                                                                        <a href="/ProdukHukum/update/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-success"><i class="fa fa-pencil"></i></a>
+                                                                        <a href="/ProdukHukum/delete/<?= md5($val['id_produk']) ?>" class="btn btn-sm btn-danger tombol-hapus"><i class="fa fa-trash"></i></a>
+                                                                    <?php } ?>
+                                                                </center>
+                                                            </td>
+                                                        </tr>
+                                                    <?php } ?>
+                                                </tbody>
+                                            </table>
+                                            <?php if ($akun['role_id'] == 3) { ?>
+                                                <button type="submit" class="btn btn-primary mt-5"><i class="fa fa-check-square-o"></i> Validasi</button>
+                                            <?php } ?>
+                                        </form>
+                                    </div>
                                 </div>
-                            </div>
+                            <?php } ?>
                         </div>
                     </div>
                 </div>
@@ -242,5 +263,42 @@
             $('.modal-body form').attr('action', '/Export/pdf_prohum');
         });
     })
+
+    $('.ganti-status').on('change', function() {
+        // alert($(this).data('id'));
+        // $('#activate').attr('disabled');
+
+        const id = $(this).data('id');
+
+        $.ajax({
+            url: '/ProdukHukum/ganti_status',
+            data: {
+                id: id
+            },
+            method: 'post',
+            dataType: 'json',
+            beforeSend: function() {
+                $($(this).data('id')).attr('disabled');
+            },
+            success: function(response) {
+                $($(this).data('id')).removeAttr('disabled');
+                if (response.berlaku) {
+                    new PNotify({
+                        title: "Berlaku",
+                        text: "Status produk hukum berlaku",
+                        type: 'success',
+                        styling: 'bootstrap3'
+                    });
+                } else {
+                    new PNotify({
+                        title: "Tidak Berlaku",
+                        text: "Status produk hukum tidak berlaku",
+                        type: 'success',
+                        styling: 'bootstrap3'
+                    });
+                }
+            }
+        });
+    });
 </script>
 <?= $this->endSection('content'); ?>
